Refactored keyword container, added table container
