+++ src/js/components/keyword/KeywordPage.jsx
@@ -8,31 +8,20 @@ import PropTypes from 'prop-types';
 
 import * as MetaTagHelper from 'helpers/metaTagHelper';
 import { InfoCircle } from 'components/sharedComponents/icons/Icons';
+import ResultsTableContainer from 'containers/keyword/table/ResultsTableContainer';
 import MetaTags from '../sharedComponents/metaTags/MetaTags';
 import Header from '../sharedComponents/header/Header';
 import Footer from '../sharedComponents/Footer';
-import ResultsTableSection from './table/ResultsTableSection';
 import KeywordHeader from './header/KeywordHeader';
 import KeywordSearchBar from './KeywordSearchBar';
 import KeywordSearchHover from './KeywordSearchHover';
 
 const propTypes = {
     updateKeyword: PropTypes.func,
-    keywordApplied: PropTypes.bool,
+    keyword: PropTypes.string,
     summary: PropTypes.object,
     summaryInFlight: PropTypes.bool,
-    error: PropTypes.bool,
-    inFlight: PropTypes.bool,
-    results: PropTypes.array,
-    columns: PropTypes.object,
-    counts: PropTypes.object,
-    sort: PropTypes.object,
-    tableTypes: PropTypes.array,
-    currentType: PropTypes.string,
-    tableInstance: PropTypes.string,
-    switchTab: PropTypes.func,
-    updateSort: PropTypes.func,
-    loadNextPage: PropTypes.func
+    fetchSummary: PropTypes.func
 };
 
 export default class KeywordPage extends React.Component {
@@ -77,7 +66,7 @@ export default class KeywordPage extends React.Component {
                     <div className="keyword-content">
                         <div className="search-bar-section">
                             <KeywordSearchBar
-                                submitText={this.props.updateKeyword} />
+                                updateKeyword={this.props.updateKeyword} />
                             <div className="info-text">
                                 Use the Keyword Search to get a broad picture of award data on a given theme.
                                 You can search through only award descriptions, or award descriptions plus other
@@ -94,20 +83,9 @@ export default class KeywordPage extends React.Component {
                                 whose extensive filters let you find more precise data sets.
                             </div>
                         </div>
-                        <ResultsTableSection
-                            inFlight={this.props.inFlight}
-                            error={this.props.error}
-                            keywordApplied={this.props.keywordApplied}
-                            tableTypes={this.props.tableTypes}
-                            currentType={this.props.currentType}
-                            switchTab={this.props.switchTab}
-                            results={this.props.results}
-                            columns={this.props.columns}
-                            counts={this.props.counts}
-                            sort={this.props.sort}
-                            updateSort={this.props.updateSort}
-                            tableInstance={this.props.tableInstance}
-                            loadNextPage={this.props.loadNextPage} />
+                        <ResultsTableContainer
+                            keyword={this.props.keyword}
+                            fetchSummary={this.props.fetchSummary} />
                     </div>
                 </main>
                 <Footer />
